Handle `CKPurchaseController.shared().perform(…)` callback within `SSPurchase.perform(adamId:purchase:)` to encapsulate `CKPurchaseController` & its callback.
Chain `Promise`s instead of calling `Promise.wait()`.

Improve comments.

Resolve #562
